requirements.txt
aiohttp
python-telegram-bot==20.7
openai>=1.0.0
google-api-python-client
google-auth-oauthlib